+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+- Show `measurementUnit` if it's different from the default.
+- Handle `unitMultiplier`.
+
 ## [1.3.2] - 2020-04-06
 ### Fixed
 - `product-summary-quantity` not working.